ad_testing/local_astrometry/requirements.txt
flask==3.0.0
astrometry
numpy
opencv-python-headless
event-stream
pillow
gunicorn
